requirements.txt
discord==0.0.2
bs4==0.0.1
requests==2.19.1
Pillow==5.3.0
dropbox==9.1.0